# backend/openapi-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AI Fight Club API",
    "description": "Complete API documentation for the AI Fight Club application - a platform for AI coding challenges and evaluation",
    "version": "1.0.0",
    "license": {
      "name": "ISC",
      "url": "https://opensource.org/licenses/ISC"
    },
    "contact": {
      "name": "Support",
      "url": "https://aifightclub.example.com",
      "email": "support@aifightclub.example.com"
    },
    "termsOfService": "https://aifightclub.example.com/terms",
    "x-api-versioning": {
      "strategy": "uri-path",
      "current": "v1",
      "deprecated": [],
      "sunset": []
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Current server (relative path)"
    }
  ],
...
  "paths": {
...
    "/evaluations": {
      "get": {
        "summary": "Get all evaluations",
        "description": "Retrieves a list of all evaluations with pagination",
        "operationId": "getAllEvaluations",
        "tags": [
          "Evaluations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/limitParam"
          },
          {
            "$ref": "#/components/parameters/offsetParam"
          },
          {
            "$ref": "#/components/parameters/sortParam"
          },
          {
            "$ref": "#/components/parameters/filterParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Evaluations retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Evaluation"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/PaginationParams"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      },
      "post": {
        "summary": "Create a new evaluation",
        "description": "Creates a new evaluation for a challenge",
        "operationId": "createEvaluation",
        "tags": [
          "Evaluations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "challengeId",
                  "userId",
                  "response"
                ],
                "properties": {
                  "challengeId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "ID of the challenge being evaluated"
                  },
                  "userId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "ID of the user who submitted the response"
                  },
                  "response": {
                    "type": "string",
                    "description": "User's response to evaluate"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Evaluation created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Evaluation"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/evaluations/{id}": {
      "get": {
        "summary": "Get evaluation by ID",
        "description": "Retrieves an evaluation by its unique identifier",
        "operationId": "getEvaluationById",
        "tags": [
          "Evaluations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the evaluation to retrieve",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Evaluation found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Evaluation"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token authentication. Use the token from the login response. Tokens expire after 24 hours."
      }
    },
    "parameters": {
      "limitParam": {
        "name": "limit",
        "in": "query",
        "description": "Maximum number of items to return per page (default: 20, max: 100)",
        "schema": {
          "type": "integer",
          "default": 20,
          "minimum": 1,
          "maximum": 100
        }
      },
      "offsetParam": {
        "name": "offset",
        "in": "query",
        "description": "Number of items to skip (for pagination)",
        "schema": {
          "type": "integer",
          "default": 0,
          "minimum": 0
        }
      },
      "sortParam": {
        "name": "sort",
        "in": "query",
        "description": "Field to sort by (prefix with - for descending order)",
        "schema": {
          "type": "string",
          "example": "-createdAt"
        }
      },
      "filterParam": {
        "name": "filter",
        "in": "query",
        "description": "Filter criteria in format field:operator:value (e.g., difficulty:eq:advanced)",
        "schema": {
          "type": "string"
        }
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Authentication information is missing or invalid",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "Unauthorized - Invalid or missing token"
                }
              }
            }
          }
        }
      },
      "NotFoundError": {
        "description": "The specified resource was not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "Resource not found"
                }
              }
            }
          }
        }
      },
      "ValidationError": {
        "description": "The request data failed validation",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "Validation error"
                },
                "errors": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "field": {
                        "type": "string"
                      },
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "schemas": {
      "PaginationParams": {
        "type": "object",
        "description": "Parameters for pagination",
        "properties": {
          "limit": {
            "type": "integer",
            "default": 20,
            "minimum": 1,
            "maximum": 100,
            "description": "Maximum number of items to return"
          },
          "offset": {
            "type": "integer",
            "default": 0,
            "minimum": 0,
            "description": "Number of items to skip"
          },
          "sort": {
            "type": "string",
            "description": "Field to sort by (prefix with - for descending order)"
          }
        }
      },
...
      "Evaluation": {
        "type": "object",
        "description": "Evaluation of a challenge response",
        "required": [
          "id",
          "challengeId",
          "userId",
          "status",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the evaluation"
          },
          "challengeId": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the challenge being evaluated"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the user who submitted the response"
          },
          "response": {
            "type": "string",
            "description": "User's response to the challenge"
          },
          "feedback": {
            "type": "string",
            "description": "Feedback on the user's response"
          },
          "score": {
            "type": "number",
            "format": "float",
            "minimum": 0,
            "maximum": 100,
            "description": "Score for the response (0-100)"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "in_progress",
              "completed",
              "failed"
            ],
            "description": "Current status of the evaluation"
          },
          "metrics": {
            "type": "object",
            "description": "Various metrics from the evaluation",
            "properties": {
              "accuracy": {
                "type": "number",
                "format": "float",
                "minimum": 0,
                "maximum": 100,
                "description": "Accuracy score"
              },
              "creativity": {
                "type": "number",
                "format": "float",
                "minimum": 0,
                "maximum": 100,
                "description": "Creativity score"
              },
              "reasoning": {
                "type": "number",
                "format": "float",
                "minimum": 0,
                "maximum": 100,
                "description": "Reasoning score"
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the evaluation was created"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the evaluation was completed"
          }
        }
      },
...
    }
  }
}